backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-multipart
scikit-learn
onnxruntime
python-Levenshtein
alembic
skl2onnx
onnx
pillow
pyzbar
requests
